refactor(pipeline): move build_comment_samples into receipts.py
The samples builder and its diagnostics join the candidate selection
and the pool in the receipts module, ahead of the verdict consumer
that grows it (#91 step 3 PR B). Pure move: outputs unchanged, tests
relocated alongside.
