+++ pipeline/receipts.py
@@ -1,20 +1,21 @@
 """The receipts: which comments stand for a player's sentiment.
 
-Candidate selection shared by the shipped comment_samples file and the
-target-verifier pool, so the two are the same computation by
-construction: the samples apply the classifier's target gate; the pool
-lifts it so named and unnamed rows compete for verification in one
-ranking.
+Home of the comment_samples fact subset and the target-verifier pool.
+Both start from one candidate selection, so they are the same
+computation by construction: the samples apply the classifier's target
+gate; the pool lifts it so named and unnamed rows compete for
+verification in one ranking.
 """
 
 import logging
 
 import polars as pl
 
-from pipeline.schemas import TARGET_POOL_SCHEMA
+from pipeline.schemas import COMMENT_SAMPLES_SCHEMA, TARGET_POOL_SCHEMA
 from utils.constants import (
     COMMENT_SAMPLES_MAX_BODY_CHARS,
     COMMENT_SAMPLES_MIN_CONFIDENCE,
+    COMMENT_SAMPLES_TOP_N,
     TARGET_POOL_SEED,
     TARGET_POOL_STRATA,
     TARGET_POOL_STRATUM_N,
@@ -157,3 +158,72 @@ def build_target_pool(
         )
     )
     return pool
+
+
+def build_comment_samples(
+    df: pl.DataFrame,
+    *,
+    n: int = COMMENT_SAMPLES_TOP_N,
+    min_confidence: float = COMMENT_SAMPLES_MIN_CONFIDENCE,
+    max_body_chars: int = COMMENT_SAMPLES_MAX_BODY_CHARS,
+) -> pl.DataFrame:
+    """
+    Select the comment samples: top-N receipts per player x sentiment.
+
+    Candidacy and ranking are select_receipt_candidates with the target
+    gate on: a polar row with no stated target is the ambiguity class a
+    receipt can't carry, while neutral rows are exempt from both polar
+    gates (the classifier reports a conventional 0.5 for neu and
+    routinely omits the target there). Thin cells are never padded.
+    Bodies are verbatim.
+
+    Args:
+        df: Attributed, flair-resolved frame with attributed_player,
+            sentiment, sentiment_player, comment_id, link_id, body,
+            score, confidence, created_utc, team.
+        n: Maximum rows per (attributed_player, sentiment) cell.
+        min_confidence: Candidacy floor on confidence, pos/neg rows only.
+        max_body_chars: Candidacy cap on body length, in characters.
+
+    Returns:
+        Frame conforming to COMMENT_SAMPLES_SCHEMA, sorted by
+        (attributed_player, sentiment, rank).
+    """
+    return (
+        select_receipt_candidates(
+            df, n=n, min_confidence=min_confidence, max_body_chars=max_body_chars
+        )
+        .rename({"team": "fan_team"})
+        .select(COMMENT_SAMPLES_SCHEMA.names())
+        .sort([*CELL, "rank"])
+    )
+
+
+def log_comment_samples_diagnostics(
+    df_attributed: pl.DataFrame, comment_samples: pl.DataFrame
+) -> None:
+    """
+    Log the multi-mention share of the sampled rows.
+
+    A two-name receipt can read ambiguously under one player's card;
+    the share is logged every run so it stays visible.
+
+    Args:
+        df_attributed: The attributed frame (carries mentioned_players).
+        comment_samples: The selected samples (COMMENT_SAMPLES_SCHEMA).
+    """
+    if not comment_samples.height:
+        logger.info("comment_samples: no rows selected")
+        return
+    # Semi-join: can't fan out if a comment_id were ever duplicated
+    multi = comment_samples.join(
+        df_attributed.filter(pl.col("mentioned_players").list.len() > 1).select(
+            "comment_id"
+        ),
+        on="comment_id",
+        how="semi",
+    ).height
+    logger.info(
+        f"comment_samples: {comment_samples.height:,} rows selected; "
+        f"{multi:,} multi-mention ({multi / comment_samples.height:.1%})"
+    )
